third_party/blink/web_tests/external/wpt/html/semantics/forms/form-submission-0/form-double-submit-requestsubmit-expected.txt

This is a testharness.js-based test.
[FAIL] test runTest2 with submitterType: input, callRequestSubmit: true, callSubmit: true, preventDefault: false, passSubmitter: true
  assert_equals: expected (object) null but got (string) "v3"
[FAIL] test runTest2 with submitterType: input, callRequestSubmit: false, callSubmit: true, preventDefault: false, passSubmitter: true
  assert_equals: expected (object) null but got (string) "v3"
[FAIL] test runTest2 with submitterType: button, callRequestSubmit: true, callSubmit: true, preventDefault: false, passSubmitter: true
  assert_equals: expected (object) null but got (string) "v3"
[FAIL] test runTest2 with submitterType: button, callRequestSubmit: false, callSubmit: true, preventDefault: false, passSubmitter: true
  assert_equals: expected (object) null but got (string) "v3"
Harness: the test ran to completion.
